Dedup code, known strings
