+++ .vscode/settings.json
@@ -75,6 +75,7 @@
         "ffmc",
         "fireweather",
         "firezone",
+        "FWIs",
         "GDPS",
         "geoalchemy",
         "GEOGCS",
@@ -85,6 +86,7 @@
         "grib",
         "gribs",
         "HAINES",
+        "Hasher",
         "hourlies",
         "HRDPS",
         "idir",
