+++ web/src/features/moreCast2/components/forecastCell.test.tsx
@@ -2,32 +2,19 @@ import { render } from '@testing-library/react'
 import ForecastCell from 'features/moreCast2/components/ForecastCell'
 import { GridRenderCellParams } from '@mui/x-data-grid-pro'
 import { vi } from 'vitest'
-import morecastInputValidSlice, { initialState } from '@/features/moreCast2/slices/validInputSlice'
-import { combineReducers, configureStore } from '@reduxjs/toolkit'
+import { initialState } from '@/features/moreCast2/slices/validInputSlice'
 import { Provider } from 'react-redux'
+import { buildTestStore } from '@/features/moreCast2/components/testHelper.test'
 
 const params: Pick<GridRenderCellParams, 'row' | 'formattedValue'> = {
   row: undefined,
   formattedValue: '1'
 }
 
-const buildTestStore = () => {
-  const rootReducer = combineReducers({
-    morecastInputValid: morecastInputValidSlice
-  })
-  const testStore = configureStore({
-    reducer: rootReducer,
-    preloadedState: {
-      morecastInputValid: initialState
-    }
-  })
-  return testStore
-}
-
 describe('ForecastCell', () => {
   it('should have input disabled when disabled prop is true', () => {
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={true}
           label="foo"
@@ -44,7 +31,7 @@ describe('ForecastCell', () => {
   })
   it('should have input enabled when disabled prop is false', () => {
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -61,7 +48,7 @@ describe('ForecastCell', () => {
   })
   it('should show less than icon when showLessThan is true', () => {
     const { queryByTestId } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -80,7 +67,7 @@ describe('ForecastCell', () => {
     const consoleErrorFn = vi.spyOn(console, 'error').mockImplementation(() => vi.fn())
     expect(() => {
       render(
-        <Provider store={buildTestStore()}>
+        <Provider store={buildTestStore(initialState)}>
           <ForecastCell
             disabled={false}
             label="foo"
@@ -95,7 +82,7 @@ describe('ForecastCell', () => {
   })
   it('should not show less than icon when showLessThan is false', () => {
     const { queryByTestId } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -111,7 +98,7 @@ describe('ForecastCell', () => {
   })
   it('should show greater than icon when showGreaterThan is true', () => {
     const { queryByTestId } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -127,7 +114,7 @@ describe('ForecastCell', () => {
   })
   it('should not show less than icon when showLessThan is false', () => {
     const { queryByTestId } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -143,7 +130,7 @@ describe('ForecastCell', () => {
   })
   it('should not show less than or greater than icons when showLessThan and showGreater than are both false', () => {
     const { queryByTestId } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -161,7 +148,7 @@ describe('ForecastCell', () => {
   })
   it('should not show a label when none specified', () => {
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label=""
@@ -177,7 +164,7 @@ describe('ForecastCell', () => {
   })
   it('should show a label when specified', () => {
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -193,7 +180,7 @@ describe('ForecastCell', () => {
   })
   it('should display the value when provided', () => {
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
@@ -214,7 +201,7 @@ describe('ForecastCell', () => {
       formattedValue: undefined
     }
     const { container } = render(
-      <Provider store={buildTestStore()}>
+      <Provider store={buildTestStore(initialState)}>
         <ForecastCell
           disabled={false}
           label="foo"
